+++ Views/MVTableView.m
@@ -76,7 +76,7 @@
 	return defaultRect;
 }
 
-- (NSRect) originalRectOfRow:(int) row {
+- (NSRect) originalRectOfRow:(NSInteger) row {
 	return [super rectOfRow:row];
 }
